Split datasets testing file into several subfiles
Aim to reduce single file testing time cost
